move constants to own folder
