+++ src/js/constants/app_constants.js
@@ -1,0 +1,5 @@
+// temporary hard-coding
+export let TARGET = 10000000;
+
+// Set to equal current fiscal year
+export var FISCAL_YEAR = '26';
