+++ src/js/constants/excel_constants.js
@@ -1,0 +1,21 @@
+// sheets to expect on detail sheet
+export const SHEETS = {
+    'FTE, Salary-Wage, & Benefits' : 'personnel' ,
+    'Overtime & Other Personnel' : 'overtime',
+    'Non-Personnel Operating' : 'nonpersonnel',
+    'Revenue' : 'revenue'
+}
+
+export const OBJ_CATEGORIES = {
+    list : [
+        // 'Salaries & Wages',
+        // 'Employee Benefits',
+        'Professional & Contractual Services',
+        'Operating Supplies',
+        'Operating Services',
+        'Equipment Acquisition',
+        'Capital Outlays',
+        'Fixed Charges',
+        'Other Expenses'
+    ]
+}
