+++ src/js/views/07_new_initiatives.js
@@ -3,7 +3,7 @@ import { View, ViewTable } from './view_class.js'
 import Table from "../components/table/table.js";
 import Form from "../components/form/form.js";
 import FundLookupTable from '../models/fund_lookup_table.js';
-import { FISCAL_YEAR } from "../init.js";
+import { FISCAL_YEAR } from '../constants/';
 
 
 const dropdownOptions = ['N/A', 'One-Time', 'Recurring']
